Moved common consts and types
