+++ src/components/map/map.tsx
@@ -1,8 +1,8 @@
 import Leaflet from "leaflet";
 import React from "react";
 
-import {MapIconUrl, MapIconSize, MapLayer, MapType} from "../const";
-import {Offer} from "../types";
+import {MapIconUrl, MapIconSize, MapLayer, MapType} from "../../const";
+import {Offer} from "../../types";
 
 import "leaflet/dist/leaflet.css";
