+++ src/store/data/types.ts
@@ -1,5 +1,5 @@
 import {ActionType} from "../const";
-import {Offer, Review} from "../../components/types";
+import {Offer, Review} from "../../types";
 
 export type DataState = {
   allOffers: Offer[],
